error func; libs recombined; files recombined; CG fixed
