+++ psarev.alexandr/T3/internalFuncs.cpp
@@ -26,40 +26,19 @@ size_t psarev::getNumVerts(Polygon& poly)
   return poly.points.size();
 }
 
-size_t psarev::getOddVerts(size_t summed, Polygon& poly)
+bool psarev::checkOddVerts(Polygon& poly)
 {
-  if (poly.points.size() % 2 != 0)
-  {
-    return ++summed;
-  }
-  else
-  {
-    return summed;
-  }
+  return poly.points.size() % 2 != 0;
 }
 
-size_t psarev::getEvenVerts(size_t summed, Polygon& poly)
+bool psarev::checkEvenVerts(Polygon& poly)
 {
-  if (poly.points.size() % 2 == 0)
-  {
-    return ++summed;
-  }
-  else
-  {
-    return summed;
-  }
+  return poly.points.size() % 2 == 0;
 }
 
-size_t psarev::getSpecVerts(size_t summed, Polygon& poly, size_t& numVerts)
+bool psarev::checkSpecVerts(Polygon& poly, size_t& numVerts)
 {
-  if (poly.points.size() == numVerts)
-  {
-    return ++summed;
-  }
-  else
-  {
-    return summed;
-  }
+  return poly.points.size() == numVerts;
 }
 
 bool psarev::isRect(Polygon& poly)
