+++ psarev.alexandr/T3/internalFuncs.hpp
@@ -8,9 +8,9 @@ namespace psarev
   int getInterValue(Point& ptOne, Point& ptTwo);
   size_t getNumVerts(Polygon& poly);
 
-  size_t getOddVerts(size_t summed, Polygon& poly);
-  size_t getEvenVerts(size_t summed, Polygon& poly);
-  size_t getSpecVerts(size_t summed, Polygon& poly, size_t& numVerts);
+  bool checkOddVerts(Polygon& poly);
+  bool checkEvenVerts(Polygon& poly);
+  bool checkSpecVerts(Polygon& poly, size_t& numVerts);
 
   bool isRect(Polygon& poly);
