+++ psarev.alexandr/T3/polygon.cpp
@@ -1,6 +1,6 @@
-#include <iostream>
 #include "polygon.hpp"
-#include "scopeGuard.hpp"
+#include <delim.hpp>
+#include <scopeGuard.hpp>
 
 using delim = psarev::DelimiterIO;
 std::istream& psarev::operator>>(std::istream& in, Polygon& data)
@@ -39,32 +39,4 @@ std::istream& psarev::operator>>(std::istream& in, Polygon& data)
     in.setstate(std::ios::failbit);
   }
   return in;
-}
-
-std::istream& psarev::operator>>(std::istream& in, DelimiterIO&& dest)
-{
-  std::istream::sentry sentry(in);
-  if (!sentry)
-  {
-    return in;
-  }
-  char c = '0';
-  in >> c;
-  if (in && (c != dest.exp))
-  {
-    in.setstate(std::ios::failbit);
-  }
-  return in;
-}
-
-using delim = psarev::DelimiterIO;
-
-std::istream& psarev::operator>>(std::istream& in, intIO&& dest)
-{
-  std::istream::sentry sentry(in);
-  if (!sentry)
-  {
-    return in;
-  }
-  return in >> dest.ref;
 }
