+++ psarev.alexandr/T3/polygon.hpp
@@ -1,24 +1,10 @@
 #ifndef POLYGON_HPP
 #define POLYGON_HPP
 #include <vector>
-#include <iostream>
+#include <ios>
 
 namespace psarev
 {
-  struct DelimiterIO
-  {
-    char exp;
-  };
-
-  std::istream& operator>>(std::istream& in, DelimiterIO&& dest);
-
-  struct intIO
-  {
-    int& ref;
-  };
-
-  std::istream& operator>>(std::istream& in, intIO&& dest);
-
   struct Point
   {
     int x, y;
