+++ psarev.alexandr/common/delim.cpp
@@ -15,3 +15,15 @@ std::istream& psarev::operator>>(std::istream& in, DelimiterIO&& dest)
   }
   return in;
 }
+
+using delim = psarev::DelimiterIO;
+
+std::istream& psarev::operator>>(std::istream& in, intIO&& dest)
+{
+  std::istream::sentry sentry(in);
+  if (!sentry)
+  {
+    return in;
+  }
+  return in >> dest.ref;
+}
