+++ psarev.alexandr/common/delim.hpp
@@ -10,6 +10,13 @@ namespace psarev
   };
 
   std::istream& operator>>(std::istream& in, DelimiterIO&& dest);
+
+  struct intIO
+  {
+    int& ref;
+  };
+
+  std::istream& operator>>(std::istream& in, intIO&& dest);
 }
 
 #endif
